requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
streamlit
langchain
langchain-community
python-dotenv